% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse.kickoff.r
\name{parse.kickoff}
\alias{parse.kickoff}
\title{Check whether this play is a kick off and if so, interpret it accordingly.}
\usage{
parse.kickoff(pbp, play)
}
\description{
Check whether this play is a kick off and if so, interpret it accordingly.
}

