% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse.plays.r
\name{parse.url}
\alias{parse.url}
\title{Scrape a game's play-by-play data}
\usage{
parse.url(url)
}
\arguments{
\item{url}{url of the website containing the data to parse, provided as a string}
}
\value{
A \code{data.table} representing each play of the game as one row.
}
\description{
Connects the the given \code{url}, scrapes the play-by-play data, and interprets it as football plays.
}
\examples{
plays = parse.url("http://espn.go.com/ncf/playbyplay?gameId=400610325&period=0")
}

