% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pbp.r
\docType{package}
\name{roxygen2}
\alias{roxygen2}
\alias{roxygen2-package}
\title{Scrape the Web for college football play-by-play data}
\description{
\package{pbp} allows you to scrape the ESPN website
for college football play-by-play data, which you can then analyze in R.
}
\details{
The only function you're likely to need from \pkg{pbp} is
\code{\link{ParsePlays}}.
}

